backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
openai==1.0.0
python-dotenv==0.21.1
